+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Added support for notion-enhancer (via @fharper)
 - Added support for GitFox (via @L3K0V)
 - Updated support for Bartender through SetApp (via @dbhagen)
+- Added support for glob pattern in application config files (via @jneuendorf
+  and @semkagtn)
 
 ## Mackup 0.8.33